feat: refactor sharing functionality and remove unused hooks
